move functionality from types to core
